+++ dashboard/15-final/app/dashboard/invoices/[id]/edit/page.tsx
@@ -1,7 +1,7 @@
 import InvoiceForm from '@/app/ui/invoices/form';
 import { notFound } from 'next/navigation';
 import { Invoice } from '@/app/lib/definitions';
-import { fetchInvoiceById } from '@/app/lib/data-fetches';
+import { fetchInvoiceById } from '@/app/lib/data';
 
 export default async function Page({ params }: { params: { id: string } }) {
   const id = params.id ? parseInt(params.id) : null;
