+++ dashboard/15-final/app/ui/customers/table.tsx
@@ -3,7 +3,7 @@ import {
   calculateCustomerInvoices,
 } from '@/app/lib/calculations';
 import { Customer, Invoice } from '@/app/lib/definitions';
-import { fetchAllCustomers, fetchAllInvoices } from '@/app/lib/data-fetches';
+import { fetchAllCustomers, fetchAllInvoices } from '@/app/lib/data';
 import Image from 'next/image';
 
 export default async function CustomersTable() {
